requirements.txt
appnope==0.1.0
backcall==0.1.0
bcrypt==3.1.7
cffi==1.12.3
decorator==4.4.0
Django==1.10
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
python-dateutil==2.8.0
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
